% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{Atkinson_fun}
\alias{Atkinson_fun}
\title{Atkinson function}
\usage{
Atkinson_fun(x, gamma = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{gamma}{A real number}
}
\value{
Value of Atkinson function \eqn{g_\gamma(x)}.
}
\description{
\code{Atkinson_fun} computes Atkinson function. It is also known as constant marginal utility of consumption function and has interpretation as prioritarian aggregating function.
}
\details{
Atkinson function is defined for any value of parameter \code{gamma}, however for negative values the function is convex and cannot be used as a prioritarian aggregating function.

\code{x} is expected to be a utility profile, i.e., none of its elements is negative. If \code{x} contains negative values, a warning message is displayed, but the value of the function is returned.

Formula of Atkinson function:
\deqn{
g_\gamma(x) =
 \left\{
   \begin{array}{ll}
     \sum_{i=1}^N \ln x_i &  \qquad \mathrm{for} \quad \gamma = 1\\
     (1-\gamma)^{-1} \sum_{i=1}^N x_i^{1-\gamma} &  \qquad \mathrm{for} \quad \gamma \neq 1
   \end{array}
 \right.
}
where \eqn{N} is the length of \eqn{x}.

The parameter \eqn{\gamma} determines the social aversion to inequality: \eqn{\gamma = 0} corresponds to no aversion (utilitarianism), \eqn{\gamma = \infty} corresponds to full aversion to inequality (egalitarianism). The parameter \eqn{\gamma} can also be interpreted as socially tolerable degree of losses involved in transfers of utility from higher levels to lower levels.
}
\examples{
Atkinson_fun(c(1, 2, 3, 4))
Atkinson_fun(c(1, 2, 3, 4), gamma = 0.5)


}
