% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{generalized_mean}
\alias{generalized_mean}
\title{Generalized mean}
\usage{
generalized_mean(x, p = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{p}{A real number}
}
\value{
Value of generalized mean \eqn{g_p(x)}.
}
\description{
\code{generalized_mean} computes the generalized mean of a vector.
}
\details{
\code{x} is expected to be a utility profile, i.e., none of its elements is negative. If \code{x} contains negative values, a warning message is displayed, but the value of the function is returned.

Formula for generalized mean:
\deqn{
g_p(x) =
 \left\{
   \begin{array}{ll}
     \left( \prod_{i=1}^N x_i \right)^{\frac{1}{N}} &  \qquad \mathrm{for} \quad p = 0\\
     \left(\frac{1}{N} \sum_{i=1}^N x_i^p \right)^{\frac{1}{p}} &  \qquad \mathrm{for} \quad p \neq 0
   \end{array}
 \right.
}
where \eqn{N} is the length of \eqn{x}.

The parameter \eqn{p} determines the trade-off between equality (egalitarianism for \eqn{p = -\infty}) and efficiency (utilitarianism for \eqn{p = 1}). For values of \eqn{p > 1} generalized mean is a convex function, attaining the higher values the more unequal is the level of inequality in a utility profile \eqn{x}, provided the total utility \eqn{\sum_{i=1}^N x_i} is kept constant.
}
\examples{
generalized_mean(c(1,2,3,4))
generalized_mean(c(1,2,3,4), alpha = 1)


}
