% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log-sum-exp function}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Value of \eqn{g(x)}.
}
\description{
\code{log_sum_exp} function is a "soft min" of a vector (i.e., a differentiable function of a vector whose value is predominantly determined by its smallest element).
}
\details{
Formula of log-sum-exp function:
\deqn{
g(x) = -\ln\left(\sum_{i=1}^{N} e^{-x_i} \right)
}
}
\examples{
log_sum_exp(c(1, 2, 3, 4))


}
