% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{minimum}
\alias{minimum}
\title{Minimum}
\usage{
minimum(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Smallest element of the vector \code{x}.
}
\description{
\code{minimum} is the \eqn{\min(\cdot)} function.
}
\details{
Formula:
\deqn{
g(x) = \min_{i = 1,\ldots,N} x_i
}
where \eqn{N} is the length of vector \code{x}.
}
\examples{
minimum(c(1, 2, 3, 4))


}
