% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{negative_Atkinson_index}
\alias{negative_Atkinson_index}
\title{Negative Atkinson index}
\usage{
negative_Atkinson_index(x, gamma = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{gamma}{A non-negative real number}
}
\value{
Value of negative Atkinson index \eqn{-I_\gamma(x)}.
}
\description{
\code{negative_Atkinson_index} returns the value of the Atkinson inequality index times -1. The higher value of negative Atkinson index the lower inequality.
}
\details{
\code{x} is expected to be a utility profile, i.e., none of its elements is negative. If \code{x} contains negative values, a warning message is displayed, but the value of the function is returned.

\code{gamma} must be a non-negative real number. If negative value of \code{gamma} is passed, a warning message is displayed and execution of the function stopped.

Formula for negative Atkinson index:
\deqn{
-I_\gamma(x) =
 \left\{
   \begin{array}{ll}
     \left(\frac{N}{\sum_{i=1}^N x_i}\right) \left( \prod_{i=1}^N x_i \right)^{\frac{1}{N}} - 1 &  \qquad \mathrm{for} \quad \gamma = 1\\
     \left(\frac{N}{\sum_{i=1}^N x_i}\right) \left( \frac{\sum_{i=1}^N x_i^{1-\gamma}}{N} \right)^{\frac{1}{1-\gamma}} - 1 &  \qquad \mathrm{for} \quad \gamma \geq 0,  \gamma \neq 1
   \end{array}
 \right.
}
where \eqn{N} is the length of \eqn{x}.

The parameter \eqn{\gamma} determines the social aversion to inequality: \eqn{\gamma = 0} corresponds to no aversion, \eqn{\gamma = \infty} corresponds to full aversion (no tolerance) to inequality.
}
\examples{
negative_Atkinson_index(c(1,2,3,4))
negative_Atkinson_index(c(1,2,3,4), alpha = 2)
}
