% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{negative_distance_to_reference_point}
\alias{negative_distance_to_reference_point}
\title{Negative distance to a reference point}
\usage{
negative_distance_to_reference_point(x, A = 0, p = 2)
}
\arguments{
\item{x}{A numeric vector}

\item{A}{A real number}

\item{p}{A real number}
}
\value{
Value of \eqn{g_{p, A}(x)}.
}
\description{
\code{negative_distance_to_reference_point} computes negative of distance in \eqn{L^p} norm between given vector and a reference point whose all coordinates are equal to a chosen aspiration level.
}
\details{
Parameter \code{p} is the exponent defining the \eqn{L^p}-norm distance. The distance is properly defined if \eqn{p \geq 1}. If \eqn{p < 1} a warning message is displayed, but the value of the function is returned.

Formula for negative distance to reference point:
\deqn{
g_{p, A}(x) = - \left( \sum_{i=1}^N |x_i - A|^p \right)^{\frac{1}{p}}
}
where \eqn{N} is the length of vector \eqn{x}.

Parameter \eqn{A < \infty} is an aspiration level of utility. Parameter \eqn{p} defines the type of \eqn{L^p} distance and also determines the trade-off between equality (egaliatarianism for \eqn{p = \infty}) and efficiency (utilitarianism for \eqn{p = 1}).
}
\examples{
negative_distance_to_reference_point(c(1,2,3,4))
negative_distance_to_reference_point(c(1,2,3,4), A = 100, p = 1.5)


}
