% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{negative_exp_sum}
\alias{negative_exp_sum}
\title{Negative exponential function}
\usage{
negative_exp_sum(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Value of \eqn{g(x)}.
}
\description{
\code{negative_exp_fun} computes negative of sum of negative exponents of elements of a vector.
}
\details{
Formula of negative exponential function:
\deqn{
g(x) = -\sum_{i=1}^N e^{-x_i}
}
where \eqn{N} is the length of vector \eqn{x}.

Negative exponential function is used as a prioritarian aggregating function that gives a very high priority to the smallest element of a utility vector \eqn{x}.
}
\examples{
negative_exp_sum(c(1, 2, 3, 4))


}
