% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{negative_generalized_entropy_index}
\alias{negative_generalized_entropy_index}
\title{Negative of generalized entropy index}
\usage{
negative_generalized_entropy_index(x, alpha = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{alpha}{A real number}
}
\value{
Value of negative generalized entropy index \eqn{-g_\alpha(x)}.
}
\description{
\code{negative_generalized_entropy_index} returns the value of generalized index times -1. The higher value of negative generalized entropy index the lower inequality.
}
\details{
\code{x} is expected to be a utility profile, i.e., none of its elements is negative. If \code{x} contains negative values, a warning message is displayed, but the value of the function is returned.

Formula for negative generalized entropy index:
\deqn{
-g_\alpha(x) =
 \left\{
   \begin{array}{ll}
     \frac{1}{N} \sum_{i=1}^N \ln{\frac{x_i}{\mu(x)}} &  \qquad \mathrm{for} \quad \alpha = 0\\
     -\frac{1}{N} \sum_{i=1}^N \frac{x_i}{\mu(x)} \ln{\frac{x_i}{\mu(x)}} &  \qquad \mathrm{for} \quad \alpha = 1\\
     \frac{-1}{N \alpha (\alpha - 1)} \sum_{i=1}^N \left( (\frac{x_i}{\mu(x)})^\alpha -1 \right) &  \qquad \mathrm{for} \quad \alpha \neq 0, 1
   \end{array}
 \right.
}
where \eqn{\mu(x)} is the mean of \code{x} and \eqn{N} is the length of \eqn{x}.

The parameter \eqn{\alpha} determines the sensitivity to changes in different segments of the utility profile \eqn{x}: lower values imply higher sensitivity to changes in the lower tail of the utility profile, while higher values imply higher sensitivity to changes in the upper tail.
}
\examples{
negative_generalized_entropy_index(c(1,2,3,4))
negative_generalized_entropy_index(c(1,2,3,4), alpha = 2)


}
