% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{negative_Gini_index}
\alias{negative_Gini_index}
\title{Negative of Gini index}
\usage{
negative_Gini_index(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Value of negative Gini index \eqn{-G(x)}.
}
\description{
\code{negative_Gini_index} returns the value of Gini index times -1. The higher value of negative Gini index the lower inequality.
}
\details{
\code{x} is expected to be a utility profile, i.e., none of its elements is negative. If \code{x} contains negative values, a warning message is displayed, but the value of the function is returned.

Formula for negative Gini index:
\deqn{
-G(x) = -\frac{\sum_{i=1}^N \sum_{i=1}^N |x_i - x_j|}{2N\sum_{j=1}^N x_j}.
}

For a proper utility profile \eqn{x}, Gini index \eqn{G} takes value between 0 and \eqn{(1-1/N)}, where \eqn{N} is the length of \eqn{x}. Value 0 corresponds to the state of perfect equality (i.e., all elements of \eqn{x} are equal), while value \eqn{(1-1/N)} is attained in the state of extreme inequality (i.e., all but one element of \eqn{x} are equal to zero).
}
\examples{
negative_Gini_index(c(1,2,3,4))
negative_Gini_index(c(1,0,0,0))
negative_Gini_index(c(1,1,1,1))
}
