% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{negative_max_downside_semideviation}
\alias{negative_max_downside_semideviation}
\title{Negative maximium downside semideviation}
\usage{
negative_max_downside_semideviation(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Value of \eqn{g(x)}.
}
\description{
\code{negative_max_downside_semideviation} computes negative of the difference between mean of a vector and its smallest element.
}
\details{
Formula:
\deqn{
g(x) = -\max_{i=1,\ldots,N} (\mu(x) - x_i)
}
where \eqn{N} is the length of vector \eqn{x} and \eqn{\mu(x) = \frac{1}{N}\sum_{i=1}^N x_i}.
}
\examples{
negative_max_downside_semideviation(c(1,2,3,4))


}
