% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{negative_mean_downside_semideviation}
\alias{negative_mean_downside_semideviation}
\title{Negative mean downside semideviation}
\usage{
negative_mean_downside_semideviation(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Value of \eqn{g(x)}.
}
\description{
\code{negative_mean_downside_semideviation} computes negative of the mean of differences between the mean of a vector and each of its elements that is smaller than the mean.
}
\details{
Formula:
\deqn{
g(x) = -\frac{1}{N} \sum_{i=1}^N (\mu(x) - x_i)_+
}
where \eqn{N} is the length of vector \eqn{x}, \eqn{\mu(x) = \frac{1}{N}\sum_{i=1}^N x_i} and \eqn{(\cdot)_+ = \max\{\cdot, 0\}}.
}
\examples{
negative_mean_downside_semideviation(c(1,2,3,4))


}
