% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{negative_total_downside_difference}
\alias{negative_total_downside_difference}
\title{Negative total downside difference}
\usage{
negative_total_downside_difference(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Value of \eqn{g(x)}.
}
\description{
\code{negative_total_downside_difference} computes negative of the sum of differences between each element of a vector and all other elements that are greater than it.
}
\details{
Formula:
\deqn{
g(x) = - \sum_{i=1}^N \sum_{j=1}^N (x_j - x_i)_+
}
where \eqn{N} is the length of vector \eqn{x} and \eqn{(\cdot)_+ = \max\{\cdot, 0\}}.
}
\examples{
negative_total_downside_difference(c(1,2,3,4))


}
