% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{OWA}
\alias{OWA}
\title{Ordered weighted average (OWA)}
\usage{
OWA(x, w, normalize_weights = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{w}{A numeric vector}

\item{normalize_weights}{Bool (TRUE by default)}
}
\value{
Value of \eqn{g_w(x)}
}
\description{
OWA computes a weighted average of a vector that was sorted in an increasing order.
}
\details{
All weights in the vector \code{w} should be non-negative for OWA to have a proper interpretation. In case of some weights being negative a warning message is displayed, but the value of the function is returned.

Formula of OWA function:
\deqn{
g_w(x) = \sum_{i=1}^N w_i \ x_{(i)}
}
where \eqn{N} is the length of vector \eqn{x} and \eqn{x_{(i)}} denotes the \eqn{i}-th smallest element of \eqn{x}.

Weights \eqn{w} forming a decreasing sequence of positive numbers, i.e., \eqn{w_1 > \ldots > w_N > 0} represent a decreasing priority to higher levels of utility in a utility profile \eqn{x}
}
\examples{
OWA(c(1, 2, 3, 4), w = c(5, 4, 3, 2), normalize_weights = FALSE)


}
