% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{type_2_achievement_fun}
\alias{type_2_achievement_fun}
\title{Type-2 achievement function}
\usage{
type_2_achievement_fun(x, A = 0, alpha = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{A}{A real number}

\item{alpha}{A real number}
}
\value{
Value of \eqn{g_{\alpha, A}(x)}
}
\description{
\code{type_2_achievement_fun} computes a composition of minimum of elements of a vector (possibly in reference to a selected aspiration level) and the sum of elements of the vector.
}
\details{
Parameter \code{alpha} should be non-negative, otherwise the function loses its interpretation. For negative values of \code{alpha} a warning message is displayed, but the value of the function is returned.

Formula of Type-2 achievement function:
\deqn{
g_{\alpha, A}(x) = \min_{i=1,\ldots,N}(x_i - A) + \alpha \sum_{i=1}^N (x_i - A)
}
where \eqn{N} is the length of vector \eqn{x}.

Type-2 achievement function is equivalent to a weighted average of vector \eqn{x} where its smallest element has a preferential weight of \eqn{(1 + 1/\alpha)} and all remaining elements have weight 1. Parameter \eqn{\alpha} determines the preference for equality (with \eqn{\alpha = 0} corresponding to strict egalitarianism and \eqn{\alpha = \infty} to utilitarianinsm).

Parameter \eqn{A} represents an aspiration level of utility.
}
\examples{
type_2_achievement_fun(c(1, 2, 3, 4))
type_2_achievement_fun(c(1, 2, 3, 4), A = 2, alpha = 0.2)


}
