% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_aggregating_funs.R
\name{underachievement_fun}
\alias{underachievement_fun}
\title{Underachievement function}
\usage{
underachievement_fun(x, alpha = 0, ineq_measure = sd)
}
\arguments{
\item{x}{A numeric vector}

\item{alpha}{A non-negative real number}

\item{ineq_measure}{A function (standard deviation by default)}
}
\value{
Value of underachievement function \eqn{g_\alpha(x)}.
}
\description{
\code{underachievement_fun} computes the value of underachievement function of a vector.
}
\details{
Formula for underachievement function:
\deqn{
g_\alpha(x) = \mu(x) - \alpha\rho(x)
}
where \eqn{\mu(x)} is the mean of \eqn{x} and \eqn{\rho(\cdot)} is an inequality measure.

The value of underachievement function is the average utility \eqn{\mu(x)} with penalty for inequality in utility distribution, as measured by the inequality measure \eqn{\rho(x)}. The parameter \eqn{\alpha} is the penalty weight. For no inequality penalty, i.e,  \eqn{\alpha = 0}, the underachievement function is equivalent to utilitarian total utility function. The higher the value of \eqn{\alpha} the larger emphasis on equality.
}
\examples{
underachievement_fun(c(1,2,3,4))
underachievement_fun(c(1,2,3,4), alpha = 0.2, ineq_measure = var)


}
